backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.7.1
joblib==1.4.0
mapbox==0.18.0
python-dotenv==1.0.1
scikit-learn==1.4.2     # Optional: If you're using ML models
numpy==1.26.4           # Optional: Often required with ML models